Gestion-deportiva V2/sigdep/requirements.txt
asgiref==3.6.0
boto3==1.26.137
botocore==1.29.137
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-jazzmin==2.6.0
django-multiupload==0.6.1
django-rest-authtoken==2.1.4
django-rest-knox==4.2.0
django-ses==3.5.0
django-storages==1.13.2
django-templated-mail==1.1.1
django-username-email==2.5.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
et-xmlfile==1.1.0
filelock==3.12.0
fontawesome==5.10.1.post1
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
Markdown==3.4.3
oauthlib==3.2.2
openpyxl==3.1.2
Pillow==10.0.0
platformdirs==3.3.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.22.0
XlsxWriter==3.1.2
